end of section 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -64,12 +65,14 @@ namespace AuctionService.Controllers
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
         {
             var auction = _mapper.Map<Auction>(auctionDto);
-            // TODO: add current user as seller
-            auction.Seller = "test";
+            // The below line from Program.cs enables us to use User.Identity.Name
+            // "options.TokenValidationParameters.NameClaimType = "username";"
+            auction.Seller = User.Identity.Name;
 
             _context.Auctions.Add(auction);
             var newAuction = _mapper.Map<AuctionDto>(auction);
@@ -88,6 +91,7 @@ namespace AuctionService.Controllers
                 new { auction.Id }, _mapper.Map<AuctionDto>(auction));
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
         {
@@ -96,7 +100,8 @@ namespace AuctionService.Controllers
 
             if (auction == null) return NotFound();
 
-            // TODO: check seller == username
+            // check if the person using the update endpoint is the same as the one who created the auction
+            if (auction.Seller != User.Identity.Name) return Forbid();
 
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -121,7 +126,8 @@ namespace AuctionService.Controllers
 
             if (auction == null) return NotFound();
 
-            // TODO: check seller == username
+            // check if the person using the delete endpoint is the same as the one who created the auction
+            if (auction.Seller != User.Identity.Name) return Forbid();
 
             _context.Auctions.Remove(auction);
